+++ src/java.base/share/classes/jdk/internal/foreign/abi/x64/sysv/CallArranger.java
@@ -265,7 +265,7 @@ public class CallArranger {
                         }
                         boolean useFloat = storage.type() == StorageType.VECTOR;
                         Class<?> type = SharedUtils.primitiveCarrierForSize(copy, useFloat);
-                        bindings.bufferLoad(offset, type, copy)
+                        bindings.bufferLoad(offset, type, (int) copy)
                                 .vmStore(storage, type);
                         offset += copy;
                     }
@@ -313,7 +313,7 @@ public class CallArranger {
                         boolean useFloat = storage.type() == StorageType.VECTOR;
                         Class<?> type = SharedUtils.primitiveCarrierForSize(copy, useFloat);
                         bindings.vmLoad(storage, type)
-                                .bufferStore(offset, type, copy);
+                                .bufferStore(offset, type, (int) copy);
                         offset += copy;
                     }
                 }
